Initial Commit; include OverrideAudit
This initial creation of the repository creates a packages.json file
as a custom package repository for inclusion in the default package
control channel to ease modification of package properties for
released packages by organization members.

This first commit includes an augmented entry for OverrideAudit, which
is based on the entry currently in the default channel file but with
an added homepage URL and tags as well as an explicitly stated URL for
the README to display on the Package Control website, so that it can
remain in sync with the released version.