Include a channel file for unreleased packages
This includes an extra channel file addition that allows access to all
of our unreleased packages via PackageControl by having a single item
that can be added to the package control repository list.

This requires that all packages listed here have tagged releases of
some sort; although it is possible to add a github repository directly
and install the package there without a tagged released, that does not
work here.