Bump the minimum build required for OverrideAudit
As of the forthcoming OverrideAudit 2.0.0 release, the minimum build of
Sublime has been bumped to coincide with the official release of Sublime
Text 3.2.

The build number of the stable release is a nice even 3200, but build
3197 is the last dev build prior to the release, and so that is being
used here to ensure that people on the dev channel aren't blocked from
receiving OverrideAudit 2.0.0 until the next dev build occurs.

Once the next dev build happens, we could revisit and change the builds
as appropriate, since 3200 is a nice even number to land on.