Include ProjectSpecificKeys in the package list
Include in the list of unreleased packages the ProjectSpecificKeys
package that has been the subject of a few live streams on the
channel. 

This package allows for key bindings to be added to a sublime-project
file, and will put those bindings in place whenever the project is
opened, removing them when the project is closed.