Include an entry for the ScratchBuffer package
This is a new package that is going to be the subject of an upcoming video on the YouTube
channel about how to get a package added to package control. For testing purposes in the
interim, we're adding it to this unreleased packages file; it can be removed when the actual
addition is scheduled.